backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
pydantic
pydantic-settings
pydantic[email]
requests
httpx
openai
pillow
numpy
faiss-cpu
langchain
langchain-openai
python-multipart
authlib
email-validator
